// src/booking/booking.enum.ts
export enum BOOKING_TYPE {
  NOW = 'NOW',
  LATER = 'LATER',
}

export enum VELET_TYPE {
  NORMAL = 'NORMAL',
  PRIORITY = 'PRIORITY',
}

export enum PAYMENT_TYPE {
  CASH = 'CASH',
  // ONLINE = 'ONLINE',
}

export enum BOOKING_STATUS {
  PENDING = 'PENDING',
  ACCEPTED = 'ACCEPTED',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
  DECLINED = 'DECLINED',
  STARTED = 'STARTED',
  ASSIGNED = 'ASSIGNED',
}

export enum BOOKING_ISSUE {}